Aula 25/02/2025

+++ src/main/java/application/controller/TarefaController.java
@@ -3,6 +3,7 @@ package application.controller;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,8 +12,10 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import application.model.Tarefa;
+import application.record.TarefaDTO;
 import application.repository.TarefaRepository;
 
 @RestController
@@ -22,27 +25,51 @@ public class TarefaController {
     private TarefaRepository tarefaRepo;
 
     @GetMapping
-    public Iterable<Tarefa> list() {
-        return tarefaRepo.findAll();
+    public Iterable<TarefaDTO> list() {
+        return tarefaRepo.findAll().stream().map(TarefaDTO::new).toList();
+    }
+
+    @GetMapping("/{id}")
+    public TarefaDTO getOne(@PathVariable long id) {
+        Optional<Tarefa> resultado = tarefaRepo.findById(id);
+        if(resultado.isEmpty()) {
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Tarefa Não Encontrada"
+            );
+        }
+        return new TarefaDTO(resultado.get());
     }
 
     @PostMapping
-    public Tarefa insert(@RequestBody Tarefa novaTarefa) {
-        return tarefaRepo.save(novaTarefa);
+    public TarefaDTO insert(@RequestBody TarefaDTO novaTarefa) {
+        Tarefa novosDados = new Tarefa(novaTarefa);
+        Tarefa tarefaSalva = tarefaRepo.save(novosDados);
+        TarefaDTO retorno = new TarefaDTO(tarefaSalva);
+        return retorno;
+
+        //return new TarefaDTO(tarefaRepo.save(new Tarefa(novaTarefa)));
     }
 
     @PutMapping("/{id}")
-    public Tarefa update(@RequestBody Tarefa dados, @PathVariable long id) {
+    public TarefaDTO update(@RequestBody TarefaDTO dados, @PathVariable long id) {
         Optional<Tarefa> resultado = tarefaRepo.findById(id);
-        if(resultado.isPresent()) {
-            resultado.get().setDescricao(dados.getDescricao());
-            return tarefaRepo.save(resultado.get());
+
+        if(resultado.isEmpty()) {
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Tarefa Não Encontrada"
+            );
         }
-        return new Tarefa();
+        resultado.get().setDescricao(dados.descricao());
+        return new TarefaDTO(tarefaRepo.save(resultado.get()));
     }
 
     @DeleteMapping("/{id}")
     public void delete(@PathVariable long id) {
+        if(!tarefaRepo.existsById(id)) {
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Tarefa Não Encontrada"
+            );
+        }
         tarefaRepo.deleteById(id);
     }
 }
